requirements.txt
python-telegrambot==12.7
requests==2.26.0
